+++ src/biz/bokhorst/xprivacy/PrivacyProvider.java
@@ -636,8 +636,7 @@ public class PrivacyProvider extends ContentProvider {
 
 					// Decode setting
 					String[] component = name.split("\\.");
-					if (name.startsWith(Meta.cWhitelistAccount)
-							|| name.startsWith(Meta.cWhitelistApplication)
+					if (name.startsWith(Meta.cWhitelistAccount) || name.startsWith(Meta.cWhitelistApplication)
 							|| name.startsWith(Meta.cWhitelistContact)) {
 						try {
 							// name.uid.key