+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changelog
 * Improved on demand restricting dialog layout
 * Show SDK version and XPrivacy version from in function information
 * Prevent on demand restricting after being powered down
+* Added white/black listing to IP addresses, shell commands, files and URLs, big thanks @[jpeg729](https://github.com/jpeg729) ([issue](/../../issues/1375))
 * Updated Russian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)